Setup Guard & guard-rpec
Introduces Guard and slightly reorganizes specs for better consistency.

You can start Guard by running `bundle exec guard`. This should run all
the Krikri specs and start file monitoring. Changing a given file will
trigger a run of the relevant specs.
