+++ Gemfile
@@ -9,6 +9,11 @@ gem 'rubocop', require: false
 
 gem 'database_cleaner', '~> 1.3.0', require: false
 
+group :development do
+  gem 'guard'
+  gem 'guard-rspec', require: false
+end
+
 file = File.expand_path('Gemfile',
                         ENV['ENGINE_CART_DESTINATION'] ||
                         ENV['RAILS_ROOT'] ||
